move content of EigenvaluesCorrelation to EigenvaluesProblems
remove temporary HCM walk

change Walk Length of test to 1
