+++ include/convex_bodies/correlation_matrices/corre_spectra.hpp
@@ -1,8 +1,7 @@
 #ifndef VOLESTI_CORRE_SPECTRAHEDRON_H
 #define VOLESTI_CORRE_SPECTRAHEDRON_H
 
-// #include "matrix_operations/EigenvaluesProblems.h"
-#include "matrix_operations/EigenvaluesCorrelation.h"
+#include "matrix_operations/EigenvaluesProblems.h"
 
 template <typename NT, typename MT, typename VT>
 struct Precompute {
@@ -52,9 +51,7 @@ class CorreSpectra : public Spectrahedron<Point> {
 
     PrecomputationOfValues _precomputedValues;
 
-    // EigenvaluesProblems<NT, MT, VT> EigenvaluesProblem;
-    EigenvaluesCorrelation<NT, MT, VT> _EigenvaluesProblem;
-
+    EigenvaluesProblems<NT, MT, VT> _EigenvaluesProblem;
 
     /// Constructor of correlation matrix spectrahedra
     /// \param[in] : matrix size
@@ -141,7 +138,7 @@ class CorreSpectra : public Spectrahedron<Point> {
 
     NT positiveLinearIntersection(VT const & p, VT const & v) {
         createMatricesForPositiveLinearIntersection(p, v);
-        return _EigenvaluesProblem.minPosLinearEigenvalue(-_precomputedValues.A, _precomputedValues.B,
+        return _EigenvaluesProblem.minPosLinearEigenvalue2(-_precomputedValues.A, _precomputedValues.B,
                                                                 _precomputedValues.eigenvector);
     }
